refactor: lots of refactoring
