+++ packages/auth/src/Expo-AuthSession/components/ExpoLoginButton.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import { Button } from "@equinor/mad-components";
 import * as WebBrowser from "expo-web-browser";
-import { useExpoAuthenticate } from "./useExpoAuthenticate";
-import { LoginButtonProps } from "../components";
-import "core-js/stable/atob";
+import { useExpoAuthenticate } from "../useExpoAuthenticate";
+import { LoginButtonProps } from "../../components";
 import { ResponseType } from "expo-auth-session";
+import "core-js/stable/atob";
 
 WebBrowser.maybeCompleteAuthSession();
