--- packages/auth/src/Expo-AuthSession/context/index.ts
@@ -1,1 +1,0 @@
-export * from "./ExpoAuth";
